requirements.txt
Click==7.0
dlib==19.16.0
Flask==1.0.2
Flask-SocketIO==3.1.0
gevent==1.3.7
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
python-engineio==3.0.0
python-socketio==2.1.0
scipy==1.1.0
six==1.11.0
uWSGI==2.0.17.1
Werkzeug==0.14.1
